Refactor task exceptions
