--- src/main/java/seedu/address/model/person/exceptions/DuplicateTaskException.java
@@ -1,11 +1,0 @@
-package seedu.address.model.person.exceptions;
-
-/**
- * Signals that the operation will result in duplicate Tasks (Tasks are considered duplicates if they have the same
- * title and description).
- */
-public class DuplicateTaskException extends RuntimeException {
-    public DuplicateTaskException() {
-        super("Operation would result in duplicate tasks");
-    }
-}
