--- src/main/java/seedu/address/model/person/exceptions/TaskNotFoundException.java
@@ -1,7 +1,0 @@
-package seedu.address.model.person.exceptions;
-
-/**
- * Signals that the operation is unable to find the specified person.
- */
-public class TaskNotFoundException extends RuntimeException {
-}
