+++ src/main/java/seedu/address/model/task/timeformat/TimeFormat.java
@@ -2,7 +2,7 @@ package seedu.address.model.task.timeformat;
 
 import java.time.LocalDateTime;
 
-import seedu.address.model.person.exceptions.TimeFormatException;
+import seedu.address.model.task.exceptions.TimeFormatException;
 
 public interface TimeFormat {
     LocalDateTime check(String time) throws TimeFormatException;
